+++ public/js/components/researchTopic/researchTopicCtrl.js
@@ -1,5 +1,5 @@
 angular.module("chasingProgress")
-    .controller("researchTopicCtrl", function($scope, $state, researchTopicSvc, toastr) {
+    .controller("researchTopicCtrl", function($scope, $state, researchTopicSvc, alertify) {
 
 
 
@@ -7,28 +7,34 @@ angular.module("chasingProgress")
         $scope.addResearchTopic = function() {
             researchTopicSvc.addResearchTopic($scope.topicName, $scope.topicImage, $scope.topicImageBackground)
                 .then(function(response) {
-                    $scope. researchTopics = response.data;
-                    console.log(response.data);
-                    toastr.success("Added " + response.data.topicName);
+                    $scope.researchTopics = response.data;
+                    alertify.success("Added " + response.data.topicName)
+                    $scope.topicName = "";
+                    $scope.topicImage  = "";
+                    $scope.topicImageBackground = "";
                 })
         }
 
         researchTopicSvc.getResearchTopics()
             .then(function(response) {
                 $scope.researchTopics = response.data;
-                console.log(response.data);
             })
 
         $scope.openResearchTopicPage = function(index) {
-            $state.go('researchTopicPage', {researchTopicId: index});
+            $state.go('researchTopicPage', {
+                researchTopicId: index
+            });
         }
 
         $scope.deleteTopic = function(topic) {
-            researchTopicSvc.deleteTopic(topic)
-                .then(function(response) {
-                    console.log('successfully deleted' + topic.topicName);
-                    toastr.success("Deleted " + response.data.topicName);
-                })
+            alertify.confirm("You are about to delete the " + topic.topicName + " topic and all of its notes. Sure you wanna do this?", function() {
+                researchTopicSvc.deleteTopic(topic)
+                    .then(function(response) {
+                        alertify.success('DELETED')
+                    })
+            }, function() {
+                alertify.error('Cancel')
+            });
         }
 
-});
+    });
